pyproject.toml
[project]
name = "decksage"
version = "0.1.0"
description = "Card similarity via tournament deck co-occurrence"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy<2.0.0",
    "scikit-learn>=1.3.0",
    "pecanpy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.1",
    "requests>=2.31.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.27.0",
    "diskcache>=5.6.3",
    "requests-cache>=1.2.1",
    "hishel>=0.1.3",
    "zstandard>=0.22.0",
    "sentence-transformers",
    "aim>=3.0.0",
    "qdrant-client>=1.7.0",
    "meilisearch>=0.32.0",
]

[project.scripts]
decksage-api = "ml.api.api:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",  # Fast, modern linter + formatter
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.6.1",
    "hypothesis>=6.98.0",
    "fastapi>=0.100.0",
    "httpx>=0.28.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
embeddings = [
    "gensim>=4.3.0",
]
viz = [
    "matplotlib>=3.7.0",
    "umap-learn>=0.5.4",
    "plotly>=5.17.0",
]
fast = [
    "fastnode2vec>=0.0.5",
    "numba>=0.57.0",
]
enrichment = [
    "pillow>=10.0.0",  # Image processing for vision models
    "openai>=1.0.0",   # For OpenRouter compatibility
]
gnn = [
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
]
stats = [
    "scipy>=1.12.0",  # Optional: for statistical functions (IAA, etc)
]

[tool.hatch.build]
# Only include Python ML code, exclude Go backend and massive data files
only-include = ["src/ml"]
exclude = [
    "src/backend",
    "src/frontend",
    "src/experiments",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ml"]

[tool.ruff]
# Modern Python linting best practices
line-length = 100
target-version = "py311"

# Enable comprehensive rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modernize code)
    "B",   # flake8-bugbear (find likely bugs)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]

# Ignore specific rules that conflict with project style
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function call in argument defaults
    "B904",  # Within an except clause, raise without from
]

# Exclude paths
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    "archive",
    "experiments/archived",
    "src/backend/**",
    "integration_test_tmp/**",
    "logs/**",
]

[tool.ruff.format]
# Use Black-compatible formatting
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["decksage", "ml"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.per-file-ignores]
# Test files can be less strict
"tests/**/*.py" = ["S101", "PLR2004"]
"**/test_*.py" = ["S101", "PLR2004"]
"src/ml/validation/validators/loader.py" = ["PLR0911", "PLR0912", "PLR0915"]
"**/validation/validators/loader.py" = ["PLR0911", "PLR0912", "PLR0915"]

[tool.pytest.ini_options]
testpaths = ["src/ml/tests"]
norecursedirs = [
    ".*",
    "build",
    "dist",
    "CVS",
    "_darcs",
    "{arch}",
    "*.egg",
    "venv",
    ".venv",
    "archive",
    "backups",
    "src/backend",
    "src/frontend", 
    "experiments",
    "integration_test_tmp",
    "logs",
    "notebooks",
    "__pycache__",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-q --strict-markers --tb=short --maxfail=1 --durations=15 --dist=loadscope -n auto"
markers = [
    "llm: marks tests that make actual LLM API calls (slow, requires API key)",
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "edge_cases: marks tests that test edge cases and error conditions",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
branch = true
source = ["src/ml"]
omit = [
  "src/backend/*",
  "src/frontend/*",
  "experiments/*",
  "integration_test_tmp/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 0
